+++ react/components/ExtensionPointError.tsx
@@ -5,6 +5,7 @@ interface Props {
   treePath: string
   error?: Error
   errorInfo?: ErrorInfo | null
+  operationIds: string[]
 }
 
 interface State {
@@ -30,8 +31,9 @@ class ExtensionPointError extends PureComponent<Props, State> {
   }
 
   public render() {
-    const { treePath, error, errorInfo } = this.props
+    const { treePath, error, errorInfo, operationIds } = this.props
     const { errorDetails } = this.state
+
     const componentStack = errorInfo && errorInfo.componentStack
 
     return (
@@ -47,9 +49,19 @@ class ExtensionPointError extends PureComponent<Props, State> {
           ({errorDetails ? 'hide' : 'show'} details)
         </button>
         {errorDetails && error && (
-          <pre>
-            <code className="f6">{error.stack}</code>
-          </pre>
+          <>
+            <ul className="f6 list pl0">
+              {operationIds.map(operationId => (
+                <li key={operationId}>
+                  <span>Operation ID:</span>{' '}
+                  <span className="i">{operationId}</span>
+                </li>
+              ))}
+            </ul>
+            <pre>
+              <code className="f6">{error.stack}</code>
+            </pre>
+          </>
         )}
         {errorDetails && componentStack && (
           <pre>